refactor(desktop): split workspace navigation flows
