+++ apps/desktop/src/hooks/workspaces/ui/use-right-panel-state-updater.ts
@@ -5,7 +5,7 @@ import {
 } from "react";
 import {
   reconcileRightPanelWorkspaceState,
-} from "@/lib/domain/workspaces/shell/right-panel-state";
+} from "@/lib/domain/workspaces/shell/right-panel-state-normalization";
 import type { RightPanelWorkspaceState } from "@/lib/domain/workspaces/shell/right-panel-model";
 import { rightPanelStateEqual } from "@/lib/domain/workspaces/shell/right-panel-view";
 import type { ViewerTarget } from "@/lib/domain/workspaces/viewer/viewer-target";
