+++ apps/desktop/src/lib/domain/preferences/workspace-ui/persisted-right-panel.ts
@@ -4,7 +4,7 @@ import {
   type RightPanelDurableState,
   type RightPanelMaterializedState,
 } from "@/lib/domain/workspaces/shell/right-panel-model";
-import { normalizeRightPanelMaterializedState } from "@/lib/domain/workspaces/shell/right-panel-state";
+import { normalizeRightPanelMaterializedState } from "@/lib/domain/workspaces/shell/right-panel-state-normalization";
 import { migrateLegacyRightPanelWorkspaceState } from "@/lib/domain/workspaces/shell/right-panel-migration";
 
 export function migrateLegacyRightPanelPreferences(args: {
